fix: init tempo precompiles in script runner
