+++ crates/forge/src/runner.rs
@@ -16,8 +16,8 @@ use foundry_common::{TestFunctionExt, TestFunctionKind, contracts::ContractsByAd
 use foundry_compilers::utils::canonicalized;
 use foundry_config::{Config, FuzzCorpusConfig};
 use foundry_evm::{
-    backend::Backend,
     constants::{CALLER, TEST_CONTRACT_ADDRESS},
+    core::tempo::FoundryStorageProvider,
     decode::RevertDecoder,
     executors::{
         CallResult, EvmError, Executor, ITest, RawCallResult,
@@ -35,10 +35,7 @@ use foundry_evm::{
 use itertools::Itertools;
 use proptest::test_runner::{RngAlgorithm, TestError, TestRng, TestRunner};
 use rayon::prelude::*;
-use revm::{
-    Database,
-    state::{AccountInfo, Bytecode},
-};
+use revm::state::{AccountInfo, Bytecode};
 use serde::{Deserialize, Serialize};
 use std::{
     borrow::Cow,
@@ -57,105 +54,6 @@ use tracing::Span;
 ///
 /// `address(uint160(uint256(keccak256("foundry library deployer"))))`
 pub const LIBRARY_DEPLOYER: Address = address!("0x1F95D37F27EA0dEA9C252FC09D5A6eaA97647353");
-
-/// Storage provider adapter for Foundry's backend to work with Tempo precompiles.
-///
-/// This wraps Foundry's backend to implement the `PrecompileStorageProvider` trait,
-/// enabling use of canonical Tempo initialization logic.
-struct FoundryStorageProvider<'a> {
-    backend: &'a mut Backend,
-    chain_id: u64,
-    timestamp: U256,
-    gas_used: u64,
-}
-
-impl<'a> FoundryStorageProvider<'a> {
-    fn new(backend: &'a mut Backend, chain_id: u64, timestamp: U256) -> Self {
-        Self { backend, chain_id, timestamp, gas_used: 0 }
-    }
-}
-
-impl<'a> tempo_precompiles::storage::PrecompileStorageProvider for FoundryStorageProvider<'a> {
-    fn chain_id(&self) -> u64 {
-        self.chain_id
-    }
-
-    fn timestamp(&self) -> U256 {
-        self.timestamp
-    }
-
-    fn set_code(
-        &mut self,
-        address: Address,
-        code: Bytecode,
-    ) -> Result<(), tempo_precompiles::error::TempoPrecompileError> {
-        self.backend.insert_account_info(
-            address,
-            AccountInfo { code_hash: code.hash_slow(), code: Some(code), ..Default::default() },
-        );
-        Ok(())
-    }
-
-    fn get_account_info(
-        &mut self,
-        _address: Address,
-    ) -> Result<&AccountInfo, tempo_precompiles::error::TempoPrecompileError> {
-        // Not needed for test initialization
-        Err(tempo_precompiles::error::TempoPrecompileError::Fatal(
-            "get_account_info not supported in test initialization".to_string(),
-        ))
-    }
-
-    fn sstore(
-        &mut self,
-        address: Address,
-        key: U256,
-        value: U256,
-    ) -> Result<(), tempo_precompiles::error::TempoPrecompileError> {
-        self.backend
-            .insert_account_storage(address, key, value)
-            .map_err(|e| tempo_precompiles::error::TempoPrecompileError::Fatal(e.to_string()))
-    }
-
-    fn sload(
-        &mut self,
-        address: Address,
-        key: U256,
-    ) -> Result<U256, tempo_precompiles::error::TempoPrecompileError> {
-        self.backend
-            .storage(address, key)
-            .map_err(|e| tempo_precompiles::error::TempoPrecompileError::Fatal(e.to_string()))
-    }
-
-    fn emit_event(
-        &mut self,
-        _address: Address,
-        _event: alloy_primitives::LogData,
-    ) -> Result<(), tempo_precompiles::error::TempoPrecompileError> {
-        // Events during initialization are not captured in test setup
-        // This is acceptable as initialization events aren't tested
-        Ok(())
-    }
-
-    fn deduct_gas(
-        &mut self,
-        gas: u64,
-    ) -> Result<(), tempo_precompiles::error::TempoPrecompileError> {
-        // Track gas for accounting purposes, but don't enforce limits during init
-        self.gas_used = self.gas_used.saturating_add(gas);
-        Ok(())
-    }
-
-    fn gas_used(&self) -> u64 {
-        self.gas_used
-    }
-
-    fn beneficiary(&self) -> Address {
-        // note(onbjerg): this doesn't matter during initialization so we can safely set it to
-        // address zero. during execution the evm will set this to an appropriate value.
-        Address::ZERO
-    }
-}
 
 /// A type that executes all tests of a contract
 pub struct ContractRunner<'a> {
